update model and reproduce to follow the repo's structure
